tests/submission/PhilipTaylor/canvas/initial.colour.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: initial.colour</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.initial.html">initial</a>.colour</h1>
<p class="desc">Initial state is transparent black</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.initial.colour">initial.colour</a>
</li></ul>
</div>
<p class="notes">Output should be transparent black (not transparent anything-else), but manual
verification can only confirm that it's transparent - it's not possible to make
the actual blackness visible.

</p><p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="initial.colour.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Initial state is transparent black");
_addTest(function(canvas, ctx) {

_assertPixel(canvas, 20,20, 0,0,0,0, "20,20", "0,0,0,0");


});
</script>

<div id="log"></div>
</body></html>
